updated files

+++ assets/back/delete_question.php
@@ -18,9 +18,33 @@ if ($conn->connect_error)
 $rawDeleteRequest = file_get_contents('php://input');
 $data = json_decode($rawDeleteRequest, true);
 $deleteRequest = array('questionID' => $data['questionID']);
+$deleteRequest = array('questionID' => 57);
 
+$existsInExam=False;
+$examRecords=mysqli_query($conn,"SELECT * FROM BETA_exams");
+if($examRecords->num_rows!=0)
+{
+  while($row=$examRecords->fetch_assoc())
+  {
+    foreach(explode(',',$row['questionIDs']) as $q)
+    {
+      if((int)$deleteRequest['questionID']==(int)$q)
+      {
+        $existsInExam=True;
+      }
+    }
+  }
+}
 
-$delete = mysqli_query($conn, "DELETE FROM BETA_questionBank WHERE questionID='".$deleteRequest['questionID']."'");
+if(mysqli_query($conn,"SELECT * FROM BETA_rawExamData WHERE questionID='".$deleteRequest['questionID']."'")->num_rows!=0 || $existsInExam==True)
+{
+  $archieve=mysqli_query($conn,"UPDATE BETA_questionBank SET archived=True WHERE questionID='".$deleteRequest['questionID']."'");
+}
+else
+{
+  $delete = mysqli_query($conn, "DELETE FROM BETA_questionBank WHERE questionID='".$deleteRequest['questionID']."'");
+}
+
 
 
 ?>
